refactor(agent-gui): tighten conversation package exports
