+++ packages/agent/gui/agent-gui/agentGuiNode/controller/agentGuiController.reporting.ts
@@ -6,7 +6,6 @@ import type {
 } from "@tutti-os/agent-activity-core";
 import { toast } from "@tutti-os/ui-system";
 import { type AgentActivityRuntime } from "../../../agentActivityRuntime";
-import type { AgentConversationBatchDeletionCapability } from "../../../agentConversationRailRuntime";
 import type { AgentHostToastApi } from "../../../host/agentHostApi";
 import type { AgentConversationVM } from "../../../shared/agentConversation/contracts/agentConversationVM";
 import type { AgentSessionState } from "../../../shared/agentSessionTypes";
@@ -27,6 +26,7 @@ import {
   getAgentGUIErrorCode,
   normalizeAgentGUIDiagnosticError
 } from "./agentGuiController.errors";
+import type { AgentConversationBatchDeletionCapability } from "./agentConversationBatchDeletionCapability";
 export {
   normalizePermissionModeSemantic,
   permissionConfigFromComposerOptions,
